Implementacion del metodo get del detalle, modificacion de las clases y creacion de los equals y hashcode

+++ src/AppForSEII2526.API/Controllers/DevicesController.cs
@@ -69,7 +69,7 @@ namespace AppForSEII2526.API.Controllers
                 m.Color,
                 m.Year,
                 m.Model.NameModel,
-                (decimal)m.PriceForRent
+                m.PriceForRent
                 ))
                 .ToListAsync();
             return Ok(devices);
